Add 3-main.c to test adding nodes at the end of a doubly linked list
This main file tests the add_dnodeint_end function by adding multiple nodes
to the end of a doubly linked list. It verifies that the function works
correctly by printing the entire list after adding the nodes. The code
follows proper structure and demonstrates how to use the function effectively